Updated test bundles to add second device
In functional testing of the set-weight action, it was found that
changes to ring devices is not triggering rebalance because there
are no changes to the partition placement.  Adding a second device to
the swift-storage units will change the partitions assigned to the disks
and allow for the rebalance and sync of rings to be validated during
functional testing.

Change-Id: I3f72da96edd97a479e0600f35d450fdf8aa077ae
Related-Bug: #1882250